Rebase branch tornado-254 onto dev
