+++ src/entities/user/api/delete-user/deleteUser.generated.ts
