+++ src/entities/user/api/delete-user/deleteUser.gql
