+++ src/entities/user/api/get-user-profile/getUserProfile.gql
