+++ src/entities/user/api/get-user-profile/getUserProfileRSC.ts
