+++ src/entities/user/index.ts
@@ -1,1 +1,17 @@
+export {
+  Delete_UserDocument,
+  type Delete_UserMutationResult,
+  type Delete_UserMutationVariables,
+  useDelete_UserMutation,
+} from './api/delete-user/deleteUser.generated'
 export { useGet_FollowersQuery } from './api/get-user-followers/userFollowers.generated'
+export {
+  Get_User_ProfileDocument,
+  type Get_User_ProfileQuery,
+  type Get_User_ProfileQueryResult,
+  type Get_User_ProfileQueryVariables,
+  useGet_User_ProfileQuery,
+  useGet_User_ProfileSuspenseQuery,
+} from './api/get-user-profile/getUserProfile.generated'
+export { type MutateUserType } from './model/domain.types'
+export { DeleteUserMenuItem } from './ui/DeleteUserMenuItem'
