+++ src/entities/user/model/domain.types.ts
