+++ src/entities/user/ui/DeleteUserMenuItem.tsx
