--- src/entities/users/index.ts
@@ -1,16 +1,0 @@
-export {
-  Delete_UserDocument,
-  type Delete_UserMutationResult,
-  type Delete_UserMutationVariables,
-  useDelete_UserMutation,
-} from './api/delete-user/deleteUser.generated'
-export {
-  Get_User_ProfileDocument,
-  type Get_User_ProfileQuery,
-  type Get_User_ProfileQueryResult,
-  type Get_User_ProfileQueryVariables,
-  useGet_User_ProfileQuery,
-  useGet_User_ProfileSuspenseQuery,
-} from './api/get-user-profile/getUserProfile.generated'
-export { type MutateUserType } from './model/domain.types'
-export { DeleteUserMenuItem } from './ui/DeleteUserMenuItem'
