--- src/pages_layer/user-profile/index.ts
@@ -1,1 +1,0 @@
-export { UserPosts } from './tabs/ui/UserPosts'
