+++ src/pages_layer/user-profile/posts/index.ts
@@ -1,0 +1,1 @@
+export { UserPosts } from './ui/UserPosts'
