+++ src/pages_layer/users-list/lib/useUserMutations.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 
-import { type MutateUserType, useDelete_UserMutation } from '@/src/entities/users'
+import { type MutateUserType, useDelete_UserMutation } from '@/src/entities/user'
 import { type GetUsersListQuery } from '@/src/widgets/users-list/table'
 
 type UseUserDeleteMutationType = {
